requirements.txt
openai
playwright
jinja2
python-dotenv
pydantic
